+++ demo/dashboard/client/src/pages/files/file/Projects/index.js
@@ -15,23 +15,16 @@ class Projects extends React.Component {
     super();
 
     this.state = {
-      budgetItemsCount: 0,
-      budgetItemsEnrichedCount: 0,
       current: 1,
-      enrichmentReportsMessage: '',
       enrichmentResults: [],
-      enrichmentReportsLoading: true,
       isLoading: false,
       pagerLength: 0,
-      locationsCount: 0,
-      locationsEnrichedCount: 0,
       projects: [],
       projectsEnriched: 0,
       total: 0,
     };
 
     this.emptyResults = this.emptyResults.bind(this);
-    this.getEnrichmentReport = this.getEnrichmentReport.bind(this);
     this.handlePageChange = this.handlePageChange.bind(this);
     this.loadData = this.loadData.bind(this);
     this.setResults = this.setResults.bind(this);
@@ -39,110 +32,23 @@ class Projects extends React.Component {
 
   componentDidMount() {
     this.clients = clients.Create();
-    this.getEnrichmentReport();
     this.loadData();
   }
 
   emptyResults() {
     this.setState({
+      current: 1,
+      enrichmentResults: [],
       isLoading: false,
-      enrichmentReportsLoading: true,
+      pagerLength: 0,
       projects: [],
+      projectsEnriched: 0,
       total: 0,
     });
   }
 
   handlePageChange(newPage) {
     this.setState({ current: newPage }, this.loadData);
-  }
-
-  getEnrichmentReport() {
-    (async () => {
-      const { computedKey } = this.props;
-      let budgetItemsCount = 0;
-      let budgetItemsEnrichedCount = 0;
-      let locationsCount = 0;
-      let locationsEnrichedCount = 0;
-      let pagination = 0;
-
-      const params = {
-        index: indices.projects,
-        type: 'project',
-        scroll: '10m',
-        q: `computed_key:"${computedKey}.ndjson"`,
-        body: {
-          size: 1000,
-          query: {
-            match_all: {},
-          },
-        },
-      };
-
-      // @see https://www.elastic.co/guide/en/elasticsearch/client/javascript-api/current/api-reference.html#_search
-      const results = await this.clients.public.search(params);
-
-      // eslint-disable-next-line
-      let { hits, _scroll_id } = results;
-
-      while (hits && hits.hits.length) {
-        pagination += hits.hits.length;
-        this.setState({
-          enrichmentReportsMessage: `${pagination} of ${hits.total}`,
-        });
-
-        // eslint-disable-next-line
-        hits.hits.forEach(record => {
-          const locations = record._source.project_locations;
-          const { budget } = record._source;
-
-          // Count occurences which contain useful data about locations and budget.
-
-          // Location is a nested document, so 1 project record can contain multiple sub-documents of a location sub-records.
-          const locationsInRecord = locations.length;
-          locationsCount += locationsInRecord;
-
-          // Budget enrichment happens in total_cost and eu_contrib.
-          if (budget.total_cost.value || budget.eu_contrib.value) {
-            budgetItemsCount += 1;
-          }
-
-          // Count how many items have been enriched.
-
-          // Location items are marked with a field "enriched".
-          const enrichedLocationsInRecord = locations.filter(
-            location => location.enriched
-          ).length;
-          locationsEnrichedCount += enrichedLocationsInRecord;
-
-          // Budget items are updated in-place, state before enrichment is stored in "_original" sub-field.
-          if (budget.total_cost._original) {
-            budgetItemsEnrichedCount += 1;
-          }
-          if (budget.eu_contrib._original) {
-            budgetItemsEnrichedCount += 1;
-          }
-        });
-
-        // eslint-disable-next-line
-        const next = await this.clients.public.scroll({
-          scroll_id: _scroll_id,
-          scroll: '10m',
-        });
-
-        // eslint-disable-next-line
-        hits = next.hits;
-        // eslint-disable-next-line
-        _scroll_id = next._scroll_id;
-      }
-
-      this.setState({
-        budgetItemsCount,
-        budgetItemsEnrichedCount,
-        enrichmentReportsLoading: false,
-        locationsCount,
-        locationsEnrichedCount,
-      });
-    })();
   }
 
   loadData() {
@@ -197,16 +103,9 @@ class Projects extends React.Component {
 
   render() {
     const {
-      budgetItemsCount,
-      budgetItemsEnrichedCount,
       current,
-      enrichmentReportsLoading,
-      enrichmentReportsMessage,
       enrichmentResults,
       isLoading,
-      length,
-      locationsCount,
-      locationsEnrichedCount,
       pagerLength,
       projects,
       projectsEnriched,
@@ -216,7 +115,8 @@ class Projects extends React.Component {
     if (isLoading) {
       return <Spinner />;
     }
-    if (!projects || total === 0) {
+
+    if (!projects || projects.length === 0) {
       return (
         <h1 className="ecl-heading ecl-heading--h1 ecl-u-mt-none">
           No projects found in current file.
@@ -224,42 +124,14 @@ class Projects extends React.Component {
       );
     }
 
-    const locationsData = [
-      {
-        name: 'Locations without enrichment',
-        value: locationsCount - locationsEnrichedCount,
-      },
-      {
-        name: 'Locations which have been enriched',
-        value: locationsEnrichedCount,
-      },
-    ];
-    const budgetData = [
-      {
-        name: 'Budget items without enrichment',
-        value: budgetItemsCount - budgetItemsEnrichedCount,
-      },
-      {
-        name: 'Budget items which have been enriched',
-        value: budgetItemsEnrichedCount,
-      },
-    ];
-
     return (
       <Listing
-        budgetData={budgetData}
-        budgetItemsCount={budgetItemsCount}
         current={current}
-        enrichmentReportsLoading={enrichmentReportsLoading}
-        enrichmentReportsMessage={enrichmentReportsMessage}
         enrichmentResults={enrichmentResults}
-        length={length}
-        locationsCount={locationsCount}
-        locationsData={locationsData}
         onHandlePageChange={this.handlePageChange}
         pagerLength={pagerLength}
-        total={total}
         projectsEnriched={projectsEnriched}
+        total={total}
       />
     );
   }
